### 2025-01-27 #7730 @editorialbot
[PRE REVIEW]: BenchmarkDataNLP.jl: Synthetic Data Generation for NLP Benchmarking
**Submitting author:** @mantzaris (<a href="http://orcid.org/0000-0002-0026-5725">Alexander V. Mantzaris</a>)
**Repository:** https://github.com/mantzaris/BenchmarkDataNLP.jl
**Branch with paper.md** (empty if default branch): main
**Version:** v.1.0.0
