### 2025-02-10 #7730 @jromanowska
:wave: @oxinabox, @ziotom78, @sglyon, @leios, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
