### 2025-02-10 #7730 @jromanowska
@mantzaris, please add the DOIs to the references in your paper - check the suggestions created by our bot [in the comment above](https://github.com/openjournals/joss-reviews/issues/7730#issuecomment-2647260390)
