### 2025-02-10 b7e05c2
now the paper's doi are added where possible
