+++ paper/paper.bib
@@ -15,7 +15,8 @@
   volume={679},
   pages={2--17},
   year={2017},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {https://doi.org/10.1016/j.tcs.2016.05.030}
 }
 
 
@@ -26,7 +27,8 @@
   journal={Revue Africaine de Recherche en Informatique et Math{\'e}matiques Appliqu{\'e}es},
   volume={15},
   year={2012},
-  publisher={Episciences. org}
+  publisher={Episciences. org},
+  doi={10.46298/arima.1956}
 }
 
 
@@ -44,5 +46,6 @@
   booktitle={2023 IEEE High Performance Extreme Computing Conference (HPEC)},
   pages={1--9},
   year={2023},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/HPEC58863.2023.10363447}
 }
