### 2025-02-25 #7844 @jromanowska
👋🏼 @mantzaris @oxinabox @TheCedarPrince, this is the review thread for the paper. All of our communications will happen here from now on.

```@editorialbot generate my checklist```

**We aim for reviews to be completed within about 2-4 weeks.** Please let me know if any of you require some more time. We can also use EditorialBot (our bot) to set automatic reminders if you know you'll be away for a known period.

I encourage you to **over-communicate** and let everyone know that you're on the task every now and then, instead of waiting several weeks to collect all your comments, questions, or suggestions.

_Please feel free to ping me (@jromanowska) if you have any questions/concerns._
