### 2025-03-10 #7844 @jromanowska
@TheCedarPrince - how is it going? Please, create your checklist and start with the review. I encourage everyone to _overcommunicate_ and inform here if you will use more time on a task or perhaps need to take a break.
