### 2025-03-31 #7844 @TheCedarPrince
Hi folks:

I finished my review of this work and paper; had a number of concerns that I'll enumerate as follows:

### General Checks

**Substantial scholarly effort: Does this submission meet the scope eligibility described in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort):** I was reviewing this and I noticed that the package barely seems to have about 3 months worth of work involved with the project; the latest commits have mostly been around the JOSS paper as opposed to package development. I am just a bit cautious on this per JOSS guidelines.

### Functionality

**Installation: Does installation proceed as outlined in the documentation?** Installation instructions are not present within the documentation.

**Functionality: Have the functional claims of the software been confirmed?** There were no provided outputs or examples that were able to be followed (there was one example, but somehow the formatting got mangled by Documenter.jl):

![Image](https://github.com/user-attachments/assets/5c898541-db48-4af4-9a08-2b4d832257aa)

**Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)** In the paper, it was claimed that

> This package also hopes to bring down the cost required when exploring different models and architectures that would require large amounts of energy, time and finances to train on large natural language datasets [@samsi2023words]. Risks of this nature can often be expected when very novel approaches are explored

I am not sure this was confirmed in the package -- did I miss it somewhere?

### Documentation

**Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.** See discussion in Functionality.

**Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).** See discussion in Functionality.

**Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support** I did not see this in the documentation anywhere.

### Software Paper

**State of the field: Do the authors describe how this software compares to other commonly-used packages?** I did not see any comparisons made in the paper to other state of the art software.

---

That was it for my reviews -- please let me know if you have any questions. Cheers!
