### 2025-03-31 #7844 @mantzaris
Thanks for the review @TheCedarPrince, I will get onto these corrections but before that I just want to make sure I understand what you think is needed.

> **Substantial scholarly effort: Does this submission meet the scope eligibility described in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort):** I was reviewing this and I noticed that the package barely seems to have about 3 months worth of work involved with the project; the latest commits have mostly been around the JOSS paper as opposed to package development. I am just a bit cautious on this per JOSS guidelines.

I put a lot of the commits together in terms of milestones, I was not certain that the raw number of commits carries weight on suitability. I do have plans to extend and support the repo.

> ### Functionality
>
> **Installation: Does installation proceed as outlined in the documentation?** Installation instructions are not present within the documentation.

The installation instructions although not in the *Documentation* they are on the README.md. In the readme shown on the repo root, it says in a section called "Installation" the steps to install:

- open the Julia REPL, get into package mode pressing ] and put: add https://github.com/mantzaris/BenchmarkDataNLP.jl, and after installation get out of package mode (backspace) and type using BenchmarkDataNLP
-  for development, clone the repo git clone https://github.com/mantzaris/BenchmarkDataNLP.jl, move into the repo directory cd BenchmarkDataNLP.jl, open the Julia REPL press ], dev ., exit the package mode and using BenchmarkDataNLP.jl

I can duplicate this into the Documentation if that helps or it should be there too.

> **Functionality: Have the functional claims of the software been confirmed?** There were no provided outputs or examples that were able to be followed (there was one example, but somehow the formatting got mangled by Documenter.jl):

In the section Functions of the Documentation, at the end of each function documented is an example. Output is described for what the user expects to see 'saved to disk'. The repo does not have a UI component, it is text processing and output examples are not part of the scope in the same sense as repos like Plots.jl etc.

> ![Image](https://github.com/user-attachments/assets/5c898541-db48-4af4-9a08-2b4d832257aa)

I will fix docstring, thank you. But it should be evident that it is simply a function call with some commented helpful explanations.

> **Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)** In the paper, it was claimed that
> > This package also hopes to bring down the cost required when exploring different models and architectures that would require large amounts of energy, time and finances to train on large natural language datasets [@samsi2023words]. Risks of this nature can often be expected when very novel approaches are explored
> I am not sure this was confirmed in the package -- did I miss it somewhere?

The data produced has regulated complexity based upon the user's setting of the complexity parameter. It is even described in each of the function documentation.

> ### Documentation
>
> **Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support** I did not see this in the documentation anywhere.

Although not stated explicitly the Issues button for the repo is open. Is that not sufficient? I can add something to the README or docs. Any recommendation?

> ### Software Paper
>
> **State of the field: Do the authors describe how this software compares to other commonly-used packages?** I did not see any comparisons made in the paper to other state of the art software.

I was not able to find any similar package.

> That was it for my reviews -- please let me know if you have any questions. Once these concerns have been addressed, please let me know. Cheers!

Thanks again @TheCedarPrince , I will get onto these changes soon.
