### 2025-04-01 9cc59a4
fixed the examples for the documentation, and the README.md
