+++ docs/make.jl
@@ -5,16 +5,20 @@ DocMeta.setdocmeta!(BenchmarkDataNLP, :DocTestSetup, :(using BenchmarkDataNLP); 
 
 
 makedocs(
-    modules = [BenchmarkDataNLP],
-    format =  Documenter.HTML(),
+    authors = "Alexander V. Mantzaris",
     sitename="BenchmarkDataNLP.jl",
+    format =  Documenter.HTML(;
+        canonical="https://mantzaris.github.io/BenchmarkDataNLP.jl",
+        edit_link="main",
+        assets=String[],
+    ),
     pages = [
         "Home" => "index.md",
         # Add more pages as needed
-    ],
-    authors = "Alexander V. Mantzaris",
+    ],    
 )
 
 deploydocs(
-    repo = "github.com/mantzaris/BenchmarkDataNLP.jl.git"
+    repo = "github.com/mantzaris/BenchmarkDataNLP.jl",
+    devbranch="main",
 )
