+++ src/rdf_data.jl
@@ -307,7 +307,7 @@ generate_rdf_corpus(
     max_filler = 2,
     use_context = true
 )
-    
+```    
 """
 function generate_rdf_corpus(complexity::Int, 
                                 num_paragraphs::Int; 
