+++ docs/src/index.md
@@ -33,6 +33,7 @@ A key feature of BenchmarkDataNLP.jl is its ability to parameterize the complexi
   - Polysemy: The degree to which words have multiple meanings or roles.
 
 By adjusting these parameters, users can create datasets ranging from simple, predictable structures to complex, more complex language patterns with elements of randomness or no randomness.
+A complexity of 100 should match a natural language in terms of the number of words and expansion types, and the user can regulate the amount to how much they want to stress their models.
 
 ## Output Format and File Structure
 
@@ -49,22 +50,57 @@ These files are named based on the user-defined base filename and are saved in t
 
 To generate a dataset with a complexity level of 100, consisting of 100,000 sentences for each of the training, validation, and test sets, you can use the following function call:
 
+```julia
 generate_corpus_CFG(
-    complexity       = 100,           # Controls grammar, vocab size, etc.
-    num_sentences    = 100_000,       # Number of text samples (lines) to generate for each file
+    complexity       = 20,           # Controls grammar, vocab size, etc.
+    num_sentences    = 1_000,       # Number of text samples (lines) to generate for each file
     enable_polysemy  = false,         # Toggle overlap of words among multiple roles
     output_dir       = "/home/user/Documents", # Output path for the files
     base_filename    = "MyDataset",   # Base name for output files
 )
+```
 
 This function will create three files in the specified directory: MyDataset_train.jsonl, MyDataset_valid.jsonl, and MyDataset_test.jsonl, containing the generated text samples partitioned accordingly.
+The output files .jsonl, will contain sentences such as (depending on the complexity that controls the vocabulary size etc):
+```
+{"text": "갇간각간가 갃갃갃 갊갋갌 갍갊갂갋갊간갂"}
+
+{"text": "갈갌갌갊 갌갋갈갃 갋갋각갃갉갌갌각"}
+```
 
 By leveraging BenchmarkDataNLP.jl, researchers and practitioners can efficiently create tailored datasets to rigorously evaluate and benchmark NLP models across a spectrum of linguistic complexities.
 
+
+## Contributing
+
+We welcome contributions from the community! To contribute to this project:
+
+1. **Fork the repository** and create your branch from `main`.
+2. **Write clear, modular code** and follow existing code conventions.
+3. **Document your changes** clearly in code and, if needed, in the `README.md`.
+4. **Test your changes** thoroughly.
+5. Submit a **Pull Request** with a clear description of what you’ve added or changed.
+
+---
+
+## Reporting Issues
+
+Found a bug or an issue? Help us improve by reporting it. Please:
+
+- Open a new [GitHub Issue](https://github.com/mantzaris/BenchmarkDataNLP.jl/issues).
+- Include a **clear and descriptive title**.
+- Provide steps to **reproduce the problem**, expected vs. actual behavior, and your system/environment details.
+- Add relevant **logs, screenshots, or error messages** to help diagnose the issue.
+
+---
+
+
 ## Functions
+
+```@index
+```
 
 ```@autodocs
 Modules = [BenchmarkDataNLP]
 Private = false
-Order = [:function]
 ```
