+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"#BenchmarkDataNLP.jl","page":"Home","title":"BenchmarkDataNLP.jl","text":"","category":"section"},{"location":"","page":"Home","title":"Home","text":"","category":"page"},{"location":"#Introduction","page":"Home","title":"Introduction","text":"","category":"section"},{"location":"","page":"Home","title":"Home","text":"BenchmarkDataNLP.jl is a Julia package designed to generate synthetic text corpora with adjustable complexity, facilitating the benchmarking and evaluation of Natural Language Processing (NLP) models, including Large Language Models (LLMs). By providing a controlled environment, researchers can systematically assess model performance across varying linguistic structures and complexities.","category":"page"},{"location":"#Purpose-and-Approach","page":"Home","title":"Purpose and Approach","text":"","category":"section"},{"location":"","page":"Home","title":"Home","text":"The primary goal of BenchmarkDataNLP.jl is to offer a suite of methods for creating synthetic datasets that mimic various linguistic patterns. These datasets are instrumental in:","category":"page"},{"location":"","page":"Home","title":"Home","text":"Model Evaluation: Assessing how NLP models handle different levels of language complexity.\nControlled Testing: Providing datasets with known properties to isolate and test specific model behaviors.\nResource Efficiency: Generating datasets that allow for efficient training and evaluation within manageable timeframes and computational resources.","category":"page"},{"location":"","page":"Home","title":"Home","text":"The package employs several methods to generate text, each offering a unique approach to dataset creation:","category":"page"},{"location":"","page":"Home","title":"Home","text":"Context-Free Grammars (CFGs): Utilizes CFGs to produce sentences based on predefined grammatical rules, allowing for the generation of structured and hierarchical language constructs.\nResource Description Framework (RDF) Templates: Incorporates RDF templates to structure data, facilitating the generation of text that aligns with specific semantic frameworks. This approach ensures that the generated text adheres to predefined data schemas, enhancing consistency and relevance.\nTemplates: Employs predefined templates to guide text generation, ensuring that the output follows specific structural patterns. This method allows for the creation of text with consistent formatting and organization, which is particularly useful for generating repetitive or formulaic content.\nFinite State Machines (FSMs): Implements FSMs to model the generation process as a series of states and transitions, enabling the production of text sequences that follow specific state-dependent rules. This approach is effective in capturing sequential dependencies and ensuring that the generated text adheres to desired patterns.","category":"page"},{"location":"","page":"Home","title":"Home","text":"Adjustable Complexity","category":"page"},{"location":"","page":"Home","title":"Home","text":"A key feature of BenchmarkDataNLP.jl is its ability to parameterize the complexity of the generated text. Users can control various parameters, including:","category":"page"},{"location":"","page":"Home","title":"Home","text":"Vocabulary Size: The number of unique words in the dataset.\nGrammar Rules: The number and complexity of rules used to generate sentences.\nSentence Length: The maximum length of generated sentences.\nPolysemy: The degree to which words have multiple meanings or roles.","category":"page"},{"location":"","page":"Home","title":"Home","text":"By adjusting these parameters, users can create datasets ranging from simple, predictable structures to complex, more complex language patterns with elements of randomness or no randomness.","category":"page"},{"locati
[…]
