+++ docs/Project.toml
@@ -1,2 +1,3 @@
 [deps]
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
+BenchmarkDataNLP = "473ed2bd-0a1f-44e6-8c06-93ed739d6613"
