### 2025-04-01 e553984
adding modules for the docs
