### 2025-04-01 #7844 @mantzaris
@TheCedarPrince (@jromanowska), I have addressed all the issues raised and they were valid points as the repo looks much better overall. I made the examples more evident, the installation approach on the readme is better with examples. The documentation gives more insight on the outputs and a recommendation for support/contributions is added. The code markup for the examples is checked. Please let me know if there is anything I missed or you think is needed in light of the changes.
