### 2025-04-01 #7844 @jromanowska
Thank you @TheCedarPrince for the review and @mantzaris for the answers.

@mantzaris, I would recommend that you created an example data generated by one of the tests, so that the reviewers can know that this is reproducible on their machines. You can also show some first rows/items of the generated data in the documentation, after showing the examples. This is important for the users to know what can they expect from your software.
