### 2025-04-01 #7844 @mantzaris
@jromanowska , please take a look at the readme, which has instructions for installation, simple code to run and a simple line of an example output, [readme link](https://github.com/mantzaris/BenchmarkDataNLP.jl?tab=readme-ov-file#quick-start-example). The Documentation does now have a section 'Usage Example' with an example and the showcasing of an example output, [link to documentation section](https://mantzaris.github.io/BenchmarkDataNLP.jl/stable/#Usage-Example).

Also, for reproducibility, there are the unit tests that can be run [link to tests](https://github.com/mantzaris/BenchmarkDataNLP.jl/blob/main/test/runtests.jl) which all pass as indicated by the CI badge on the repo. This covers all the functionality discussed.

Is there anything that I have missed or you believe that more of something is needed?
