### 2025-04-06 #7844 @oxinabox
This seems fine. As far as meeting JOSS requirements.

 A few comments:

### It's only in Korean.

It is under decribed in the paper and the documentation that this is only for us with Korean (Hangul).
Hungal is mentioned, but i feel like it should be mentioned far earlier.
Like in the description at the start

> BenchmarkDataNLP.jl is a package written in Julia Lang for generating synthetic text corpora, **in Korean** ...

It's unclear to me exactly how hard it would be to generalize it beyond Korean, but i do see in the text it is hard coded.
You can't pass in a separate alphabet/vocab afaict.

### Output format:

The fact that this generates a JSONNL file is a bit odd. I would have expected there to be functions for generating just a plan list of strings, and then maybe some helper functions for writing to file (though even here I would have expected the format just to be line seperated plain text).
having functions that let this be used without writing to disk would make it easier for someone else to use this as part of automated tests etc.

### Not enough control for most interesting use cases.
The most interesting use cases are exploring model failure modes, and architectures for handling specialised text patterns.
As is stated at the end of the paper.
Towards this end, I would want to be able to set things up to generate [garden path sentences](https://en.wikipedia.org/wiki/Garden-path_sentence),  sentences with multiple [center embeddings](https://en.wikipedia.org/wiki/Center_embedding),  [Zeugma/Syllepsis](https://en.wikipedia.org/wiki/Zeugma_and_syllepsis) etc.
But right now this doesn't really let me directiong constrain any of the methods as to what they will produce in ways that would let me be certain to generate any features like this.
With that said, it does produce some amazingly cursed sentence, such as [this one it made me ](https://translate.google.com.au/?sl=auto&tl=en&text=%EA%B0%8A%EA%B0%80%EA%B0%8D%EA%B0%85%20%EA%B0%86%EA%B0%88%EA%B0%8D%EA%B0%88%EA%B0%8A%20%EA%B0%88%EA%B0%88%EA%B0%86%EA%B0%86%EA%B0%87%EA%B0%85%EA%B0%89%20%EA%B0%8A%EA%B0%87%EA%B0%80%EA%B0%81%EA%B0%8C%EA%B0%87%EA%B0%82%20%EA%B0%8A%20%EA%B0%82%EA%B0%8D%EA%B0%80%20%EA%B0%81%EA%B0%86%EA%B0%82%EA%B0%87%EA%B0%82%EA%B0%8B%EA%B0%8B%20%EA%B0%84%EA%B0%86%EA%B0%80%EA%B0%89%EA%B0%85%EA%B0%87%EA%B0%8D.%20%EA%B0%87%EA%B0%80%EA%B0%8B%EA%B0%84%EA%B0%88%20%EA%B0%87%EA%B0%8C%EA%B0%82%EA%B0%87%EA%B0%8A%EA%B0%89%EA%B0%83%20%EA%B0%8B%EA%B0%89%EA%B0%88%EA%B0%80%EA%B0%82%EA%B0%88%EA%B0%8A%EA%B0%8D%20%EA%B0%85%EA%B0%89%EA%B0%84%EA%B0%86%EA%B0%85%EA%B0%87%EA%B0%87%EA%B0%81%20%EA%B0%8B%20%EA%B0%84%EA%B0%86%EA%B0%80%EA%B0%89%EA%B0%85%EA%B0%87%EA%B0%8D.%20%EA%B0%81%EA%B0%8C%EA%B0%89%EA%B0%84%EA%B0%80%EA%B0%86%20%EA%B0%88%EA%B0%80%EA%B0%84%EA%B0%87%EA%B0%8A%EA%B0%89%EA%B0%89%EA%B0%84%20%EA%B0%8A%EA%B0%84%EA%B0%84%EA%B0%89%EA%B0%82%20%EA%B0%86%EA%B0%88%EA%B0%87%EA%B0%8A%EA%B0%87%EA%B0%86%EA%B0%84%EA%B0%84%20%EA%B0%8B%20%EA%B0%8C%EA%B0%84%EA%B0%88%EA%B0%85%20%EA%B0%84%EA%B0%86%EA%B0%80%EA%B0%89%EA%B0%85%EA%B0%87%EA%B0%8D.%22%0A&op=translate). Which is cursed in content, and in how it is written and expecially in how it sounds in Korean
