### 2025-04-07 #7844 @mantzaris
> It is under decribed in the paper and the documentation that this is only for us with Korean (Hangul).

In the paper this sentence is added (also in the Docs): "The reason for this choice is that Hangul Syllables block is one of the largest continuous blocks in the Unicode standard. It contains over 11,000 characters without interruption allowing for large alphabet and punctuation subsets to be selected."

> The fact that this generates a JSONNL file is a bit odd...

That is a good point. My thinking was that I see JSONNL (.jsonl, or jsonnd) being used more frequently. JSONNL does not have the issue of becoming a big singular file that the parser has to go through all at once being monolithic. It can get streamed line by line like plain text, but can add attribute information if desired later.

> Towards this end, I would want to be able to set things up to generate garden path sentences, sentences with multiple Zeugma and syllepsis etc.

Center embeddings are a great idea, I will keep that in mind for the future. For Zeugma and syllepsis, garden path sentences, I am not sure how that can be done with the overall approach taken in the package where 'stochastic' generation guides the expansions.

> it does produce some amazingly cursed sentence, such as...

That is a stunning example!
