### 2025-04-07 6a47810
addition to the docs clarying the reason for the Hangul sequence start
