### 2025-04-15 #7844 @oxinabox
I do think the phrase "Korean language" needs to be somewhere in the first paragraph of the paper and the readme.
