### 2025-04-15 445c99f
now the paper and readme is more clear that the language production samples alphabet characters from then hangul block in unicode
