### 2025-05-16 #7844 @jromanowska
@mantzaris, could you go through the points above in **Additional Author Tasks**? :)
