### 2025-05-16 86a331b
changes to the paper references
