+++ paper/paper.bib
@@ -16,7 +16,7 @@
   pages={2--17},
   year={2017},
   publisher={Elsevier},
-  doi = {https://doi.org/10.1016/j.tcs.2016.05.030}
+  doi = {10.1016/j.tcs.2016.05.030}
 }
 
 
