### 2025-05-19 #7844 @jromanowska
Thanks! I can see that the files are on figshare, but I can't see the version of the software that is archived. And you have a different license on figshare (CC-BY) vs on github (MIT) - I think this should be fixed.

Another question: why haven't you released on github the latest version? (You have a tag with v1.0.2)
