### 2025-05-19 #7844 @jromanowska
I had commented on my pull request, but I saw that my comments were somehow gone when you merged it. I placed my comments once more in [the pull request thread](https://github.com/mantzaris/BenchmarkDataNLP.jl/pull/1).
