### 2025-05-19 5a1352d
fixes to the paper
