+++ paper/paper.bib
@@ -1,12 +1,9 @@
-
 @inproceedings{copestake1996applying,
   title={Applying natural language processing techniques to speech prostheses},
   author={Copestake, Ann},
   booktitle={Working Notes of the 1996 AAAI Fall Symposium on Developing Assistive Technology for People with Disabilities},
   year={1996}
 }
-
-
 
 @article{maletti2017survey,
   title={Survey: Finite-state technology in natural language processing},
@@ -19,8 +16,6 @@
   doi = {10.1016/j.tcs.2016.05.030}
 }
 
-
-
 @article{faye2012survey,
   title={A survey of RDF storage approaches},
   author={Faye, David C and Cure, Olivier and Blin, Guillaume},
@@ -30,7 +25,6 @@
   publisher={Episciences. org},
   doi={10.46298/arima.1956}
 }
-
 
 @article{van1996generalized,
   title={Generalized context-free grammars},
@@ -48,4 +42,21 @@
   year={2023},
   organization={IEEE},
   doi={10.1109/HPEC58863.2023.10363447}
+}
+
+@article{wang2019superglue,
+  title={Superglue: A stickier benchmark for general-purpose language understanding systems},
+  author={Wang, Alex and Pruksachatkun, Yada and Nangia, Nikita and Singh, Amanpreet and Michael, Julian and Hill, Felix and Levy, Omer and Bowman, Samuel},
+  journal={Advances in neural information processing systems},
+  volume={32},
+  year={2019},
+  doi={10.5555/3454287.3454581}
+}
+
+@article{wang2018glue,
+  title={GLUE: A multi-task benchmark and analysis platform for natural language understanding},
+  author={Wang, Alex and Singh, Amanpreet and Michael, Julian and Hill, Felix and Levy, Omer and Bowman, Samuel R},
+  journal={arXiv preprint arXiv:1804.07461},
+  year={2018},
+  doi={10.18653/v1/W18-5446}
 }
