### 2025-05-20 #7844 @mantzaris
@jromanowska , The article proof looks good to me.
