### 2025-05-21 #7844 @jromanowska
@mantzaris - please fix the DOI issue found by editorialbot
