+++ paper/paper.bib
@@ -44,13 +44,17 @@
   doi={10.1109/HPEC58863.2023.10363447}
 }
 
-@article{wang2019superglue,
-  title={Superglue: A stickier benchmark for general-purpose language understanding systems},
-  author={Wang, Alex and Pruksachatkun, Yada and Nangia, Nikita and Singh, Amanpreet and Michael, Julian and Hill, Felix and Levy, Omer and Bowman, Samuel},
-  journal={Advances in neural information processing systems},
-  volume={32},
-  year={2019},
-  doi={10.5555/3454287.3454581}
+
+
+@misc{wang2019superglue,
+      title={SuperGLUE: A Stickier Benchmark for General-Purpose Language Understanding Systems}, 
+      author={Alex Wang and Yada Pruksachatkun and Nikita Nangia and Amanpreet Singh and Julian Michael and Felix Hill and Omer Levy and Samuel R. Bowman},
+      year={2020},
+      eprint={1905.00537},
+      archivePrefix={arXiv},
+      primaryClass={cs.CL},
+      url={https://arxiv.org/abs/1905.00537}, 
+      doi={10.48550/arXiv.1905.00537}
 }
 
 @article{wang2018glue,
