### 2025-05-21 #7844 @mantzaris
@jromanowska , yes I overlooked that, sorry

- new DOI: 10.6084/m9.figshare.29119301
- new version is the released
- reference DOI issue fixed
