### 2025-05-22 #7844 @jromanowska
Wonderful! Just one more thing: could you add somewhere to figshare information about the version of the archived software? It's nowhere to be found there... Is it also possible to add your affiliation there? Or perhaps link your author name to ORCiD or something...
