### 2025-05-22 #7844 @mantzaris
@jromanowska , yes, I added the version number to the figshare title. In the author details I put my university/institution email with the orcid. It should be visible from the user account linking.
