### 2025-05-22 #7844 @jromanowska
Oh, I haven't noticed the ORCiD linking on figshare account - that's great!
But I can't see the version number [in figshare](https://figshare.com/articles/software/Benchmark_Data_NLP_BenchmarkDataNLP_jl_/29119301?file=54709472)
