### 2025-05-23 #7844 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
