### 2025-05-28 #7844 @arfon
@mantzaris – also, congratulations on being the 3000th published paper in JOSS!
